Export from snapshot service
